correction of the uint208 casting by modifying _ampletowaample function

+++ contracts/ElasticVault.sol
@@ -253,12 +253,12 @@ contract ElasticVault is AMPLRebaser, Wrapper, Ownable, ReentrancyGuard {
     */
     function makeDeposit(uint256 amount) _rebaseSynced() nonReentrant() external {
         ampl_token.safeTransferFrom(msg.sender, address(this), amount);
-        uint256 waampl = _ampleTowaample(amount);
+        uint208 waampl = _ampleTowaample(amount);
         // first deposit needs to initialize the linked list
         if(_deposits[msg.sender].nodeIdCounter == 0) {
             _deposits[msg.sender].initialize();
         }
-        _deposits[msg.sender].insertEnd(DepositsLinkedList.Deposit({amount: uint208(waampl), timestamp:uint48(block.timestamp)}));
+        _deposits[msg.sender].insertEnd(DepositsLinkedList.Deposit({amount: waampl, timestamp:uint48(block.timestamp)}));
 
         uint256 to_mint = amount.mul(10**9).divDown(EEFI_DEPOSIT_RATE);
         uint256 deposit_fee = to_mint.mul(DEPOSIT_FEE_10000).divDown(10000);
